Add encode test

+++ src/bvc_enc/bvc_enc.h
@@ -15,7 +15,7 @@ public:
 	bvc_encoder();
 
 	bvc_enc_result init(bvc_enc_config* in_config);
-	bvc_enc_result encode(bvc_picture* in_picture, bvc_enc_nal** out_nal_units, int* out_num_nal_units);
+	bvc_enc_result encode(bvc_picture* in_picture, bvc_enc_nal** out_nal_units, size_t* out_num_nal_units);
 
 private:
 	int get_size_in_bytes(bvc_chroma_format in_format);
